Restore "Android Library Project" functionality.
This is lingering cleanup in the transition from sbt 0.7 to newer versions.
It restores the "Android Library Project" functionality:  if an Android
project has other Android projects in its direct classPath dependencies,
and you build an APK for that project, it will include resources for the
base project and its dependencies, and the IDs will be made consistent.
(In fact, by generating the same R.java file for all of them, with the
appropriate package name for each; it's a kludge, but that's how the
stock library project support is working now, too.)

This differs slightly from the previous version, in that it will include
only Android projects declared as direct dependencies, and not all the
Android projects in the build.  There may be only one sbt-android build
on the planet complex enough that it makes sense to be this selective,
but it happens to be mine ;-).

(It might make sense to include resources for dependencies-of-dependencies,
but I'm not sure anyone has a practical use for that...)