Simplify TypedResources usage a bit.
This patch redefines the 'TypedViewHolder' trait so that it can be
applied to anything with a findViewById method, using structural
typing --- without requiring anything other than the 'findViewById'
method itself to exist.  The 'def view = this' and 'def activity = this'
of the TypedViewHolder and TypedActivityHolder traits are no longer
required; it's now possible to say

  class MyActivity extends Activity with TypedViewHolder {
    ...
  }

without further ceremony.

For back-compatibility, TypedActivityHolder is still defined as a
synonym for TypedViewHolder (though TypedViewHolder serves for both,
and may be better named --- a TypedActivityHolder holds Views, not
Activities).

The implicit conversions in the TypedResource object are likewise
still defined, and still work.