in fact, skip proguard by default in tests
proguard is run on the main project, which should roll in the scala library
and most other dependencies.  The test project is only going to roll in test
dependencies; by default, it's unlikely those are going to be in desperate
need of paring down, and I for one would rather not incur the time cost.